ai_employee_scripts/pyproject.toml
[project]
name = "ai-employee-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Google API (shared by watchers and MCP)
    "google-api-python-client>=2.189.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    # Watcher dependencies
    "playwright>=1.58.0",
    "watchdog>=6.0.0",
    # MCP server dependencies
    "mcp>=0.1.0",
    "playwright-stealth>=2.0.2",
    "httpx>=0.28.1",
    "odoorpc>=0.10.1",
    "tweepy>=4.16.0",
    # Cloud Agent dependencies (Platinum Tier)
    "openai-agents>=0.1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "flask>=3.1.3",
    "requests>=2.32.5",
]
